requirements.txt
# Common requirements
boto3==1.24.76
paramiko==2.11.0
scp==0.14.4
requests==2.28.1
PyYAML==6.0
psutil==5.9.2
protobuf<3.20,>=3.9.2
grpcio>=1.24.3,<2.0
fastflow-tensorflow==2.7.0

# Test
moto==4.0.5
pytest==7.1.2
pytest-cov==3.0.0
pytest-html==3.1.1
pytest-mock==3.8.2
pytest-random-order==1.0.4

# Build
grpcio-tools>=1.24.3,<1.46.0
wheel==0.37.1
setuptools==61.2.0
